+++ src/i18n/locales/en/translation.json
@@ -30,6 +30,9 @@
         }
       }
     },
+    "duration_seconds": "{{ seconds }}s",
+    "duration_minutes": "{{ minutes }}m",
+    "duration_minutes_seconds": "{{ minutes }}m {{ seconds }}s",
     "errors": {
       "error_loading_wallet_failed": "Loading wallet failed. {{ reason }}",
       "error_reloading_wallet_failed": "Reloading wallet failed. {{ reason }}",
@@ -792,16 +795,11 @@
     "progress_current_state_waiting_confirmation": "Waiting for action <1>{{ current }}</1> of <3>{{ total }}</3> ({{ txid }}) to be confirmed.",
     "progress_current_state_confirmed": "Transaction <1>{{ current }}</1> of <3>{{ total }}</3> confirmed.",
     "progress_current_state_creating_next": "Executing action <1>{{ current }}</1> of <3>{{ total }}</3>...",
-    "progress_wait_seconds": "{{ seconds }}s",
-    "progress_wait_minutes": "{{ minutes }}m",
-    "progress_wait_minutes_seconds": "{{ minutes }}m {{ seconds }}s",
-    "progress_schedule_info_title": "Schedule details",
     "progress_entry_label": "Action {{ index }}",
     "progress_entry_state_pending": "Pending",
     "progress_entry_state_waiting_confirmation": "Waiting for confirmation ({{ txid }})",
     "progress_entry_state_confirmed": "Confirmed",
-    "progress_entry_wait_before_next": "Wait before next: {{ wait }}",
-    "progress_entry_wait_final": "Final transaction",
+    "progress_entry_wait_before_next_title": "Wait period before next action",
     "progress_done": "All actions completed successfully. The scheduler will stop soon.",
     "precondition": {
       "hint_missing_utxos": "To run the scheduler you need UTXOs with {{ minConfirmations }} or more confirmations. $t(scheduler.precondition.nested_hint_fund_wallet, {\"count\": {{ minConfirmations }} })",
